+++ controllers/ImportController.php
@@ -20,7 +20,6 @@ class ImportController extends BaseController
      */
     public function actionGetEntryTypes()
     {
-
         // Only ajax post requests
         $this->requirePostRequest();
         $this->requireAjaxRequest();
@@ -41,7 +40,6 @@ class ImportController extends BaseController
      */
     public function actionUpload()
     {
-
         // Get import post
         $import = craft()->request->getRequiredPost('import');
 
@@ -63,6 +61,18 @@ class ImportController extends BaseController
             // Save file to Craft's temp folder for later use
             $file->saveAs($path);
 
+            // Get source
+            $source = craft()->assetSources->getSourceTypeById($import['assetsource']);
+
+            // Get folder to save to
+            $folderId = craft()->assets->getRootFolderBySourceId($import['assetsource']);
+
+            // Move the file by source type implementation
+            $response = $source->insertFileByPath($path, $folderId, $file->getName(), true);
+
+            // Prevent sensitive information leak. Just in case.
+            $response->deleteDataItem('filePath');
+
             // Put vars in model
             $model = new ImportModel();
             $model->filetype = $file->getType();
@@ -76,7 +86,7 @@ class ImportController extends BaseController
                 // Send variables to template and display
                 $this->renderTemplate('import/_map', array(
                     'import' => $import,
-                    'file' => $path,
+                    'file' => $response->getDataItem('fileId'),
                     'columns' => $columns,
                 ));
             } else {
@@ -96,26 +106,26 @@ class ImportController extends BaseController
      */
     public function actionImport()
     {
-
         // Get import post
         $settings = craft()->request->getRequiredPost('import');
 
         // Get file
-        $file = craft()->request->getParam('file');
+        $fileId = craft()->request->getParam('file');
+        $file = craft()->assets->getFileById($fileId);
 
         // Get mapping fields
         $map = craft()->request->getParam('fields');
         $unique = craft()->request->getParam('unique');
 
         // Get rows/steps from file
-        $rows = count(craft()->import->data($file));
+        $rows = count(craft()->import->data($file->id));
 
         // Proceed when atleast one row
         if ($rows) {
 
             // Set more settings
             $settings = array_merge(array(
-                'file' => $file,
+                'file' => $file->id,
                 'rows' => $rows,
                 'map' => $map,
                 'unique' => $unique,
@@ -130,18 +140,8 @@ class ImportController extends BaseController
             // UNCOMMENT FOR DEBUGGING
             //craft()->import->debug($settings, $history, 1);
 
-            // Determine new folder to save original importfile
-            $folder = dirname($file).'/'.$history.'/';
-            IOHelper::ensureFolderExists($folder);
-
-            // Move the file to its history folder
-            IOHelper::move($file, $folder.basename($file));
-
-            // Update the settings with the new file location
-            $settings['file'] = $folder.basename($file);
-
             // Create the import task
-            $task = craft()->tasks->createTask('Import', Craft::t('Importing').' '.basename($file), $settings);
+            $task = craft()->tasks->createTask('Import', Craft::t('Importing').' '.$file->filename, $settings);
 
             // Notify user
             craft()->userSession->setNotice(Craft::t('Import process started.'));